Made gamepad 2 off by default.

+++ src/movie.cpp
@@ -1314,7 +1314,7 @@ bool FCEUMOV_ReadState(EMUFILE* is, uint32 size)
 				//if we made it this far, then the savestate has identical movie data but we want to know now if the stateMOVIE size is greater than current movie size and make this error
 				
 				//currFrameCounter at this point represents the savestate framecount
-				if(currFrameCounter > currMovieData.records.size()) 
+				if(currFrameCounter > (int)currMovieData.records.size()) 
 				{
 					//TODO: turn frame counter to red to get attention
 					FCEU_PrintError("Savestate is from a frame (%d) after the final frame in the movie (%d). This is not permitted.", tempMovieData.records.size(), currMovieData.records.size()-1);
